:sparkles: feat: move updateLearnerProgress from jump block to process message service
